+++ fmgen/fmg_wrap.h
@@ -8,7 +8,7 @@ void OPM_Cleanup(void);
 void OPM_Reset(void);
 void OPM_Update(int16_t *buffer, int length, int rate, uint8_t *pbsp, uint8_t *pbep);
 void FASTCALL OPM_Write(uint32_t r, uint8_t v);
-uint8_t FASTCALL OPM_Read(uint16_t a);
+uint8_t FASTCALL OPM_Read(uint32_t a);
 void FASTCALL OPM_Timer(uint32_t step);
 void OPM_SetVolume(uint8_t vol);
 void OPM_SetRate(int clock, int rate);
